+++ src/utils/core/index.js
@@ -1,0 +1,14 @@
+/*
+
+	ard-eventhub
+	by SWR audio lab
+
+*/
+
+const createHashedId = require('./createHashedId')
+const getPublisher = require('./getPublisher')
+
+module.exports = {
+	createHashedId,
+	getPublisher,
+}