+++ src/utils/pubsub/createTopic.js
@@ -5,17 +5,27 @@
 
 */
 
+// load node utils
+const slug = require('slug')
+const moment = require('moment')
+
 // load pubsub for internal queues
 const publisherClient = require('./_publisherClient')
 const config = require('../../../config')
 
 module.exports = async (newTopic) => {
 	// create new topic
-	const prefix = 'projects/ard-eventhub/topics/'
 	const topic = {
-		name: prefix + newTopic.pubsub,
+		name: `projects/${process.env.GCP_PROJECT_ID}/topics/${newTopic.name}`,
 		labels: {
-			name: newTopic.label,
+			created: moment().format('YYYY-MM-DD'),
+			'creator-slug': slug(newTopic.creator),
+
+			id: newTopic.id,
+
+			'institution-slug': slug(newTopic.institution.title),
+			'publisher-slug': slug(newTopic.publisher.title),
+
 			stage: config.stage,
 		},
 	}