+++ src/utils/pubsub/getSubscriptions.js
@@ -9,14 +9,23 @@
 const pubSubClient = require('./_client')
 const mapSubscription = require('./mapSubscription')
 
+// load config
+const config = require('../../../config')
+
 module.exports = async () => {
 	// fetch subscriptions list
 	let [subscriptions] = await pubSubClient.getSubscriptions()
 
-	// DEV filter subscriptions by prefix
+	// filter subscriptions by prefix (stage)
+	subscriptions = subscriptions.filter((subscription) => subscription.name.indexOf(config.pubSubPrefix) !== -1)
 
 	// map and filter values
-	subscriptions = await Promise.all(subscriptions.map(async (subscription) => mapSubscription(subscription)))
+	subscriptions = await Promise.all(
+		subscriptions.map(async (subscription) => {
+			const { limited } = await mapSubscription(subscription)
+			return Promise.resolve(limited)
+		})
+	)
 
 	// return data
 	return Promise.resolve(subscriptions)