+++ src/utils/pubsub/getTopics.js
@@ -7,20 +7,26 @@
 
 // load pubsub for internal queues
 const pubSubClient = require('./_client')
+const convertId = require('./convertId')
+
+// load config
 const config = require('../../../config')
 
 module.exports = async () => {
 	// fetch topic list
 	let [topics] = await pubSubClient.getTopics()
 
-	// filter topics by stage
-	topics = topics.filter((topic) => topic.name.split('/').pop().indexOf(`.${config.stage}.`) !== -1)
+	// filter topics by prefix (stage)
+	topics = topics.filter((topic) => topic.name.indexOf(config.pubSubPrefix) !== -1)
 
 	// map values
 	topics = topics.map((topic) => {
+		const name = topic.name.split('/').pop()
+
 		return {
 			type: 'PUBSUB',
-			name: topic.name.split('/').pop(),
+			id: convertId.decode(name).replace(config.pubSubPrefix, ''),
+			name,
 			path: topic.name,
 			labels: topic.metadata.labels,
 		}