+++ src/utils/pubsub/index.js
@@ -5,13 +5,26 @@
 
 */
 
+const buildId = require('./buildId')
+const createSubscription = require('./createSubscription')
+const createTopic = require('./createTopic')
+const convertId = require('./convertId')
+const deleteSubscription = require('./deleteSubscription')
+const getSubscription = require('./getSubscription')
+const getSubscriptions = require('./getSubscriptions')
+const getTopic = require('./getTopic')
+const getTopics = require('./getTopics')
+const publishMessage = require('./publishMessage')
+
 module.exports = {
-	createSubscription: require('./createSubscription'),
-	deleteSubscription: require('./deleteSubscription'),
-	getSubscription: require('./getSubscription'),
-	getSubscriptions: require('./getSubscriptions'),
-	getTopic: require('./getTopic'),
-	getTopics: require('./getTopics'),
-	createTopic: require('./createTopic'),
-	publishMessage: require('./publishMessage'),
+	buildId,
+	convertId,
+	createSubscription,
+	createTopic,
+	deleteSubscription,
+	getSubscription,
+	getSubscriptions,
+	getTopic,
+	getTopics,
+	publishMessage,
 }